Fix wso2/product-ei/#2042
When custom SOAPHeader is added to a message in secured proxy service
by enrich mediator java.lang.ClassCastException: occurs since OMElement
Impl cannot be cast to SOAPHeaderBlock in rampart. Reason for this is
in enrich mediator custom SOAPHeader added as an OMElementImpl this
has been fixed by adding SOAPHeaderBlock instead.

Fixes: https://github.com/wso2/product-ei/issues/2042