Fix duplicate WS Security Headers
When a secEnabled Proxy service is calling another secEnabled Proxy to compile a response, the resulting payload will have 2 wsse:Security headers in the SOAPHeader. As per the SOAP 1.1 specification, only one Security header should be present if the role or actor attributes are not specified. Therefore, when adding the custom headers to the wsse security header, we first check and re-use any existing Security header element.

However, the custom headers should not be added to the wsse:Security header if the response is a SOAP payload. Fixed the if condition to suit this rule.

Fixes : https://github.com/wso2/product-ei/issues/2123